+++ raffia/src/parser/sass.rs
@@ -22,6 +22,7 @@ impl<'cmt, 's: 'cmt> Parser<'cmt, 's> {
     pub(super) fn parse_sass_at_rule(
         &mut self,
         at_keyword_name: &str,
+        span: Span,
     ) -> PResult<Option<(Statement<'s>, bool)>> {
         debug_assert!(matches!(self.syntax, Syntax::Scss | Syntax::Sass));
         match at_keyword_name {
@@ -62,6 +63,10 @@ impl<'cmt, 's: 'cmt> Parser<'cmt, 's> {
             "warn" => Ok(Some((Statement::SassWarnAtRule(self.parse()?), false))),
             "error" => Ok(Some((Statement::SassErrorAtRule(self.parse()?), false))),
             "debug" => Ok(Some((Statement::SassDebugAtRule(self.parse()?), false))),
+            "else" => Err(Error {
+                kind: ErrorKind::UnexpectedSassElseAtRule,
+                span,
+            }),
             _ => Ok(None),
         }
     }