+++ raffia/src/parser/stmt.rs
@@ -260,7 +260,8 @@ impl<'cmt, 's: 'cmt> Parser<'cmt, 's> {
         let mut statements = Vec::with_capacity(1);
         loop {
             let mut is_block_element = false;
-            match &peek!(self).token {
+            let TokenWithSpan { token, span } = &peek!(self);
+            match token {
                 Token::Ident(..) | Token::HashLBrace(..) | Token::AtLBraceVar(..) => {
                     if is_top_level {
                         statements.push(Statement::QualifiedRule(self.parse()?));
@@ -306,7 +307,7 @@ impl<'cmt, 's: 'cmt> Parser<'cmt, 's> {
                     Syntax::Scss | Syntax::Sass => {
                         let at_keyword_name = at_keyword.ident.name();
                         if let Some((statement, is_block)) =
-                            self.parse_sass_at_rule(&at_keyword_name)?
+                            self.parse_sass_at_rule(&at_keyword_name, span.clone())?
                         {
                             statements.push(statement);
                             is_block_element = is_block;